Move ble to #1 package structure and import RaptPillData
