+++ app/src/main/java/com/brewthings/app/data/model/RaptPillData.kt
@@ -1,0 +1,10 @@
+package com.brewthings.app.data.model
+
+data class RaptPillData(
+    val temperature: Float,
+    val gravity: Float,
+    val x: Int,
+    val y: Int,
+    val z: Int,
+    val battery: Float
+)
