+++ app/src/main/java/com/brewthings/app/di/Koin.kt
@@ -1,6 +1,6 @@
 package com.brewthings.app.di
 
-import com.brewthings.app.ble.RaptPillScanner
+import com.brewthings.app.data.ble.RaptPillScanner
 import com.brewthings.app.ui.screens.scanning.ScanningScreenViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
