+++ app/src/main/java/com/brewthings/app/ui/screens/scanning/ScanningScreen.kt
@@ -30,7 +30,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.brewthings.app.R
-import com.brewthings.app.ble.RaptPill
+import com.brewthings.app.data.model.RaptPill
 import com.brewthings.app.ui.components.ExpandableCard
 import com.brewthings.app.ui.components.ScanPane
 import com.brewthings.app.ui.theme.Typography
